refactor(pipeline): extract SampleFilter and PipelineUtils
Phase 4a of pipeline deduplication refactor (#174). Extracts
character-for-character identical filterSamples() and durationMs()
from both pipelines into shared SampleFilter and PipelineUtils enums.

72 lines deleted from both pipelines. Zero-risk warmup before Phase 3.

record_tts validated: full pipeline fires correctly (ASR, paste cascade
tier=cgevent, pipeline timing 1.060s).

Closes #178
